Rename pop-over component
