+++ src/components/pop-over/pop-over.vue
@@ -1,0 +1,8 @@
+<template>
+	<div :class="side" class="popover">
+		<slot />
+	</div>
+</template>
+
+<script src="./pop-over.js"></script>
+<style src="./pop-over.scss" scoped></style>
