--- src/components/popover/popover.vue
@@ -1,8 +1,0 @@
-<template>
-	<div :class="side" class="popover">
-		<slot />
-	</div>
-</template>
-
-<script src="./popover.js"></script>
-<style src="./popover.scss" scoped></style>
