+++ src/components/repository-line/repository-line.js
@@ -1,5 +1,5 @@
 import Badge from '../badge/badge.vue'
-import Popover from '../popover/popover.vue'
+import Popover from '@/components/pop-over/pop-over.vue'
 import RepositoryRemover from '../repository-remover/repository-remover.vue'
 import BuildStatuses from '../build-statuses/build-statuses.vue'
